requirements.txt
django==1.8.3
djangorestframework==3.1.3
requests==2.7.0
whitenoise==2.0.2
